Remove conflicts with upstream MySQL* package names
If we wanted to conflict with names from other popular repos, we'd have
to track down all those names, which include variations of mysql-*,
MySQL-*, mariadb-*, and MariaDB-*.  Not all the subpackages align
either, making it even more complicated.  The best course is to opt out
and remind people that mixing repos is a bad idea.